fix: resetting retry counter on message processing error

+++ shared/src/commonMain/kotlin/io/mrnateriver/smsproxy/shared/models/MessageEntry.kt
@@ -7,7 +7,7 @@ data class MessageEntry(
     val guid: UUID,
     val externalId: String?,
     val sendStatus: MessageRelayStatus,
-    val sendRetries: UShort,
+    val sendRetries: Int,
     val sendFailureReason: String?,
     val messageData: MessageData,
     val createdAt: Instant?,
